Review changes: Refactor test fixtures to use RegisteredLaunchDescription and clean up unused code
