+++ ros2_ws/src/rcdt_franka/test/integration/test_franka_moveit.py
@@ -6,9 +6,8 @@
 import launch_pytest
 import pytest
 from launch import LaunchDescription
-from launch.actions import IncludeLaunchDescription
-from rcdt_utilities.launch_utils import assert_for_message, assert_for_node
-from rcdt_utilities.register import Register
+from rcdt_utilities.launch_utils import assert_for_message
+from rcdt_utilities.register import Register, RegisteredLaunchDescription
 from rcdt_utilities.test_utils import get_joint_position, wait_for_register
 from rclpy.node import Node
 from sensor_msgs.msg import JointState
@@ -17,9 +16,9 @@ from utils import call_move_to_configuration_service
 
 @launch_pytest.fixture(scope="module")
 def franka_and_moveit_launch(
-    core_launch: IncludeLaunchDescription,
-    controllers_launch: IncludeLaunchDescription,
-    moveit_launch: IncludeLaunchDescription,
+    core_launch: RegisteredLaunchDescription,
+    controllers_launch: RegisteredLaunchDescription,
+    moveit_launch: RegisteredLaunchDescription,
 ) -> LaunchDescription:
     return Register.connect_context([core_launch, controllers_launch, moveit_launch])
 
@@ -32,20 +31,6 @@ def test_wait_for_register(timeout: int) -> None:
 @pytest.mark.launch(fixture=franka_and_moveit_launch)
 def test_joint_states_published(timeout: int) -> None:
     assert_for_message(JointState, "franka/joint_states", timeout=timeout)
-
-
-@pytest.mark.launch(fixture=franka_and_moveit_launch)
-def test_moveit_node_activated(test_node: Node) -> None:
-    assert_for_node(
-        fully_qualified_node_name="franka/move_group",
-        test_node=test_node,
-        timeout=30,
-    )
-    assert_for_node(
-        fully_qualified_node_name="franka/moveit_manager",
-        test_node=test_node,
-        timeout=30,
-    )
 
 
 @pytest.mark.launch(fixture=franka_and_moveit_launch)
