+++ ros2_ws/src/rcdt_panther/test/conftest.py
@@ -4,30 +4,29 @@
 
 
 import pytest
-from launch.actions import IncludeLaunchDescription
 from rcdt_utilities.launch_utils import get_file_path
-from rcdt_utilities.test_utils import *  # noqa
+from rcdt_utilities.register import RegisteredLaunchDescription
 
 
 @pytest.fixture(scope="module")
-def core_launch() -> IncludeLaunchDescription:
+def core_launch() -> RegisteredLaunchDescription:
     """Fixture to create launch file for the panther core."""
-    return IncludeLaunchDescription(
+    return RegisteredLaunchDescription(
         get_file_path("rcdt_panther", ["launch"], "core.launch.py")
     )
 
 
 @pytest.fixture(scope="module")
-def controllers_launch() -> IncludeLaunchDescription:
+def controllers_launch() -> RegisteredLaunchDescription:
     """Fixture to create launch file for the panther controllers."""
-    return IncludeLaunchDescription(
+    return RegisteredLaunchDescription(
         get_file_path("rcdt_panther", ["launch"], "controllers.launch.py")
     )
 
 
 @pytest.fixture(scope="module")
-def navigation_launch() -> IncludeLaunchDescription:
+def navigation_launch() -> RegisteredLaunchDescription:
     """Fixture to create launch file for the panther navigation."""
-    return IncludeLaunchDescription(
+    return RegisteredLaunchDescription(
         get_file_path("rcdt_panther", ["launch"], "navigation.launch.py")
     )
