+++ ros2_ws/src/rcdt_panther/test/integration/test_panther_core.py
@@ -8,8 +8,8 @@ import launch_pytest
 import pytest
 from geometry_msgs.msg import Twist
 from launch import LaunchDescription
-from launch.actions import IncludeLaunchDescription
 from rcdt_utilities.launch_utils import assert_for_message
+from rcdt_utilities.register import Register, RegisteredLaunchDescription
 from rcdt_utilities.test_utils import (
     call_trigger_service,
     get_joint_position,
@@ -21,15 +21,10 @@ from sensor_msgs.msg import JointState
 
 @launch_pytest.fixture(scope="module")
 def panther_core_launch(
-    core_launch: IncludeLaunchDescription, controllers_launch: IncludeLaunchDescription
+    core_launch: RegisteredLaunchDescription,
+    controllers_launch: RegisteredLaunchDescription,
 ) -> LaunchDescription:
-    return LaunchDescription(
-        [
-            core_launch,
-            controllers_launch,
-            launch_pytest.actions.ReadyToTest(),
-        ]
-    )
+    return Register.connect_context([core_launch, controllers_launch])
 
 
 @pytest.mark.launch(fixture=panther_core_launch)
